+++ welly/tools.py
@@ -6,25 +6,20 @@ Some extra bits.
 """
 import numpy as np
 import wellpathpy as wp
-
-from . import utils
-from .curve import Curve
+import warnings
 
 
-class RGBLog(object):
-    """
-    Attempt at RGB. Incomplete.
-    """
-    def __init__(self, curves):
-        pass
-
-
-def compute_position_log(deviation, td=None, method='mc', azimuth_datum=0):
+def compute_position_log(deviation,
+                         td=None,
+                         method='mc',
+                         azimuth_datum=0,
+                         course_length=1,
+                         ):
     """Use wellpathpy to return the required welly objects
 
     Args:
         deviation (ndarray): A deviation survey with rows like MD, INC, AZI
-        td (Number): The TD of the well, if not the end of the deviation
+        td (float): The TD of the well, if not the end of the deviation
             survey you're passing.
         method (str):
             'mc': minimum curvature
@@ -33,16 +28,16 @@ def compute_position_log(deviation, td=None, method='mc', azimuth_datum=0):
             'hi': high tangential
             'lo': low tangential
             'rc': radius of curvature
-
-        update_deviation: This function makes some adjustments to the dev-
-            iation survey, to account for the surface and TD. If you do not
-            want to change the stored deviation survey, set to False.
-        azimuth_datum (Number): The orientation of the azimuth datum,
-            relative to the y-axis.
+        azimuth_datum (float): The orientation of the azimuth datum, relative
+            to the y-axis.
+        course_length (float): Default: 1. The length over which to normalize
+            the dogleg severity. Typical values are 30 m or 100 ft. Use 1 for
+            no normalization. Note: from v0.5, default will be 30.
 
     Returns:
         ndarray. A position log with rows like X-offset, Y-offset, Z-offset
     """
+    deviation = np.asanyarray(deviation)
 
     # Adjust to TD.
     if td is not None:
@@ -57,30 +52,34 @@ def compute_position_log(deviation, td=None, method='mc', azimuth_datum=0):
     # Adjust to azimuth_datum.
     deviation += [0, 0, azimuth_datum]
 
-    # Make wellpathpy objects
-    md, inc, azi = np.split(deviation, 3, 1)
-    dev = wp.deviation(md, inc, azi)
+    # Make wellpathpy deviation object.
+    dev = wp.deviation(*deviation.T)
 
-    if method == 'mc':
-        pos = dev.minimum_curvature()
-    elif method == 'aa':
-        pos = dev.tan_method()
-    elif method == 'bt':
-        pos = dev.tan_method(choice='bal')
-    elif method == 'hi':
-        pos = dev.tan_method(choice='high')
-    elif method == 'lo':
-        pos = dev.tan_method(choice='low')
-    elif method == 'rc':
-        pos = dev.radius_curvature()
-    else:
-        msg = 'uknown choice {}, must be one of {}'
-        methods = 'mc aa bt hi lo rc'.split()
-        raise ValueError(msg.format(method, ' '.join(methods)))
+    # Compute position log.
+    methods = {
+        'mc': dev.minimum_curvature(course_length),
+        'rc': dev.radius_curvature(),
+        'aa': dev.tan_method(),
+        'bt': dev.tan_method(choice='bal'),
+        'hi': dev.tan_method(choice='high'),
+        'lo': dev.tan_method(choice='low'),
+    }
+    try:
+        pos = methods[method]
+    except KeyError:
+        msg = 'Unknown choice of method: {}. '.format(method)
+        msg += 'Method must be one of mc, aa, bt, hi, lo, or rc.'
+        raise KeyError(msg)
 
-    # Convert to welly formats
-    deviation = np.hstack((dev.md, dev.inc, dev.azi))
-    position = np.hstack((pos.easting.reshape(-1, 1), pos.northing.reshape(-1, 1), pos.depth.reshape(-1, 1)))
-    dogleg = np.deg2rad(dev.minimum_curvature().dls.reshape(-1, 1))
+    # Convert to welly formats.
+    # deviation = np.hstack((dev.md, dev.inc, dev.azi))
+    position = np.stack([pos.easting, pos.northing, pos.depth]).T
+    dogleg = dev.minimum_curvature(course_length=course_length).dls
+
+    # Keep dls in radians but warn about it. Changing in next version.
+    msg = "Dogleg severity is in radians. From v0.5 it will be returned as "
+    msg += "degrees per course length (a new argument)."
+    warnings.warn(msg, category=FutureWarning)
+    dogleg = np.radians(dogleg)
 
     return deviation, position, dogleg